정수 우선도 수정

+++ src/pyodide-simulator/models.py
@@ -13,9 +13,9 @@ class DietProcess(Process):
     # DIET 전용 프로세스: 프론트에서 받은 식탐(appetite)을 tick별 I/O interrupt 확률로 보존한다.
     appetite: int
     # DIET의 기본 스케줄링 우선순위이며, ready 대기/CPU 하차 규칙에 따라 변한다.
-    priority: float = 0.0
+    priority: int = 0
     # I/O interrupt 이후 ready queue에 복귀한 프로세스에 1틱 동안만 주는 진입 보너스다.
-    enter_bonus: float = 0.0
+    enter_bonus: int = 0
     # enter_bonus를 제거하기까지 남은 tick 수다.
     enter_bonus_ticks: int = 0
     # 4초 이상 연속 실행으로 long decay가 적용됐는지 기록해 CPU 하차 감점 여부를 결정한다.
@@ -42,9 +42,9 @@ class ProcessMetric:
 @dataclass
 class ReadyQueuePriority:
     pid: str
-    priority: float
-    enter_bonus: float
-    score: float
+    priority: int
+    enter_bonus: int
+    score: int
 
 
 @dataclass
